refactor(server): promote routes under server/ subtree
